Refactor a lot of small stuff
