+++ src/components/message-queue/message-queue.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {connect} from 'react-redux';
 import {dispatchers, stateTypes} from 'utilities';
-import MessageComponent from './MessageComponent';
+import VisualMessage from './visual-message';
 
 interface Props {
     messages: stateTypes.Message[];
@@ -11,7 +11,7 @@ function MessageQueue(props: Props) {
     return (
         <div className='fixed bottom-0 z-30 mx-0 w-100vw md:mx-20vw md:w-60vw xl:w-30vw xl:mx-35vw'>
             {props.messages.map((message: stateTypes.Message) => (
-                <MessageComponent
+                <VisualMessage
                     key={message.text}
                     message={message}
                     close={() => props.closeMessage(message.text)}
