+++ src/pages/index.ts
@@ -1,3 +1,3 @@
-export {default as FormPage} from './FormPage';
-export {default as NotFoundPage} from './NotFoundPage';
-export {default as DashboardPage} from './DashboardPage';
+export {default as FormPage} from './pages-types/form-page';
+export {default as DashboardPage} from './pages-types/dashboard-page';
+export {default as NotFoundPage} from './not-found-page';
