+++ src/pages/pages-types/dashboard-page.tsx
@@ -2,7 +2,7 @@ import React, {useEffect} from 'react';
 import {connect} from 'react-redux';
 import {stateTypes, configTypes, dispatchers, fetchSurveys} from 'utilities';
 import {Navbar} from 'components';
-import 'styles/DashboardPage.scss';
+import 'styles/dashboard-page.scss';
 
 interface Props {
     children: React.ReactNode;
